Moved initialization of the EventEnricher to the separate class
